+++ src/main.rs
@@ -1,6 +1,8 @@
 mod app;
 mod player;
 mod physics;
+mod rope;
+mod event;
 
 fn main() {
     app::run();
